+++ src/Snap/Util/Proxy.hs
@@ -22,14 +22,11 @@ module Snap.Util.Proxy
 
 ------------------------------------------------------------------------------
 import           Control.Applicative   (Alternative ((<|>)))
-import           Control.Arrow         (second)
-import qualified Data.ByteString.Char8 as S (break, breakEnd, drop, dropWhile, readInt, spanEnd)
+import           Control.Monad         (mfilter)
+import qualified Data.ByteString.Char8 as S (breakEnd, dropWhile, null, readInt, spanEnd)
 import           Data.Char             (isSpace)
-import           Data.Maybe            (fromJust)
+import           Data.Maybe            (fromMaybe)
 import           Snap.Core             (MonadSnap, Request (rqClientAddr, rqClientPort), getHeader, modifyRequest)
-#if !MIN_VERSION_base(4,8,0)
-import           Control.Applicative   ((<$>))
-#endif
 ------------------------------------------------------------------------------
 
 
@@ -81,16 +78,13 @@ xForwardedFor req = req { rqClientAddr = ip
                         , rqClientPort = port
                         }
   where
-    proxyString  = getHeader "Forwarded-For"   req <|>
-                   getHeader "X-Forwarded-For" req <|>
-                   Just (rqClientAddr req)
+    extract = fst . S.spanEnd isSpace . S.dropWhile isSpace . snd . S.breakEnd (== ',')
 
-    proxyAddr    = trim . snd . S.breakEnd (== ',') . fromJust $ proxyString
+    ip      = fromMaybe (rqClientAddr req) $ mfilter (not . S.null) $ fmap extract $
+              getHeader "Forwarded-For"   req  <|>
+              getHeader "X-Forwarded-For" req
 
-    trim         = fst . S.spanEnd isSpace . S.dropWhile isSpace
-
-    (ip,portStr) = second (S.drop 1) . S.break (== ':') $ proxyAddr
-
-    port         = fromJust (fst <$> S.readInt portStr <|>
-                             Just (rqClientPort req))
+    port    = maybe (rqClientPort req) fst $ (S.readInt =<<) $ fmap extract $
+              getHeader "Forwarded-Port"   req  <|>
+              getHeader "X-Forwarded-Port" req
 {-# INLINE xForwardedFor #-}