+++ changelog.md
@@ -1,3 +1,15 @@
+## 1.0.3.0
+### Added
+- Alternative file upload handling into Snap.Util.FileUploads
+
+### Fixes
+- Fixed parsing of field values in multipart/form-data headers with encodings
+  other than US-ASCII
+
+## 1.0.2.1
+### Dependencies
+- Allow `io-streams` 1.4.
+
 ## 1.0.2.0
 ### Added
 - Merged CORS functionality from snap-cors project into Snap.Util.CORS